requirements.txt
pytz==2021.1
python-dotenv==0.19.2
chromedriver-autoinstaller
selenium
psycopg2==2.9.1